+++ Demo/ApplozicSwiftDemo/LoginViewController.swift
@@ -39,7 +39,7 @@ class LoginViewController: UIViewController {
         }
         alUser.userId = userName.text
         ALUserDefaultsHandler.setUserId(alUser.userId)
-        print("userName:: ", alUser.userId)
+        print("userName:: ", alUser.userId ?? "")
         if !((emailId.text?.isEmpty)!) {
             alUser.email = emailId.text
             ALUserDefaultsHandler.setEmailId(alUser.email)