+++ Sources/Controllers/ALKConversationViewController.swift
@@ -1899,8 +1899,7 @@ extension ALKConversationViewController: ALMQTTConversationDelegate {
         }
     }
 
-    public func syncCall(_ alMessage: ALMessage!, andMessageList messageArray: NSMutableArray!) {
-        print("sync call1 ", messageArray)
+    public func syncCall(_ alMessage: ALMessage!, andMessageList _: NSMutableArray!) {
         guard let message = alMessage else { return }
         sync(message: message)
     }