+++ Sources/Utilities/AVCaptureVideoOrientation+Extension.swift
@@ -16,6 +16,8 @@ extension AVCaptureVideoOrientation {
         case .landscapeRight: return .landscapeRight
         case .portrait: return .portrait
         case .portraitUpsideDown: return .portraitUpsideDown
+        @unknown default:
+            return .unknown
         }
     }