+++ Sources/Utilities/Array+Extension.swift
@@ -24,7 +24,7 @@ extension Array {
 extension Array where Element: Equatable {
     // Remove first collection element that is equal to the given `object`:
     mutating func remove(object: Element) {
-        if let index = index(of: object) {
+        if let index = firstIndex(of: object) {
             remove(at: index)
         }
     }