+++ Sources/Views/ALKPhotoCell.swift
@@ -173,7 +173,7 @@ class ALKPhotoCell: ALKChatBaseCell<ALKMessageViewModel>,
         NSLog("Messages with attachment: ", messages)
 
         guard let viewModel = viewModel as? ALKMessageModel,
-            let currentIndex = messageModels.index(of: viewModel) else { return }
+            let currentIndex = messageModels.firstIndex(of: viewModel) else { return }
         vc?.viewModel = ALKMediaViewerViewModel(messages: messageModels, currentIndex: currentIndex, localizedStringFileName: localizedStringFileName)
         UIViewController.topViewController()?.present(nav!, animated: true, completion: {
             button.isEnabled = true